backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
pytest
requests
celery
redis
gunicorn